==> code/2_data-process_DIA.R <==
# 清除工作区
rm(list = ls())
# 加载必要的库
suppressWarnings(suppressMessages(library(openxlsx)))
suppressWarnings(suppressMessages(library(tidyverse)))
suppressWarnings(suppressMessages(library(impute)))
suppressWarnings(suppressMessages(library(Biostrings)))
suppressWarnings(suppressMessages(library(Peptides)))
suppressWarnings(suppressMessages(library(SeqKnn)))
# 定义文件路径和参数
conf <- "./config/default.conf"
source(conf)
uniprot_file <- Sys.glob(paste0(uniprot_file_path,organism_category,"/*",organism_code, "_", organism_ID, "_annotation.xlsx"))
fasta_file <- list.files(path = "data", pattern = "\\.fasta$", full.names = TRUE)
kegg_file <- Sys.glob(paste0(kegg_file_path,organism_category,"/",organism_code,"_KO_pathway_description.csv"))
interpro_file <- Sys.glob(paste0(interpro_file_path,organism_Scientific_name,"/*",".tsv"))
output1 <- "./result/2_Data_Summary/"
if (!dir.create(output1)) {
  dir.create(output1,recursive = T)
}
output2 <- "./result/4_Diff_Expressed/4.1_DiffStats/"
if (!dir.create(output2)) {
  dir.create(output2,recursive = T)
}
group.file <- read.xlsx("./temp/group.xlsx")
group_num <- length(unique(group.file$Group))
sample_num <- length(group.file$Group) / length(unique(group.file$Group))
nonsamplecol <- 4
group.name <- unique(group.file$Group)
contrast.file <- read.xlsx("./temp/contrast.xlsx")
compare.name <- contrast.file$contrast
foldchange1 <- Fold_Change1
foldchange2 <- Fold_Change2
dif_p <- pvalue
##读入蛋白数据
protein.data.ori <- read.delim(protein_file,header = T,check.names = F)
peptide.data.ori <- read.delim(peptide_file,header = T,check.names = F)
# 修改列名以提取样本名
process_proteins_colnames <- function(data, group_num, sample_num) {
  colnames(data)[6 : ncol(data)] <- sub(".*/([^/_]*)_.*", "\\1", colnames(data)[6 : ncol(data)])
  # colnames(data)[6 : ncol(data)] <- str_replace(str_extract(colnames(data)[6 : ncol(data)], "([A-Z0-9]+)_"),"_","")
  return(data)
}
process_peptides_colnames <- function(data, group_num, sample_num) {
  colnames(data)[11 : ncol(data)] <- sub(".*/([^/_]*)_.*", "\\1", colnames(data)[11 : ncol(data)])
  # colnames(data)[11 : ncol(data)] <- str_replace(str_extract(colnames(data)[11 : ncol(data)], "([A-Z0-9]+)_"),"_","")
  return(data)
}
protein.data <- process_proteins_colnames(protein.data.ori, group_num, sample_num)
peptide.data <- process_peptides_colnames(peptide.data.ori, group_num, sample_num)
write.xlsx(protein.data,file = paste0(output1,"ProteinGroup.xlsx"))
write.xlsx(peptide.data,file = paste0(output1,"Peptide.xlsx"))
# 处理蛋白质ID和基因列
process_ids_genes <- function(data) {
  fasta.data <- readAAStringSet(fasta_file)
  data$Protein.Accessions <- sapply(strsplit(data$Protein.Group, split = ";"), function(x) x[1])
  fasta.name <- sapply(strsplit(names(fasta.data),"\\|"), function(x) x[2])
  match.sequence <- fasta.data[fasta.name %in% data$Protein.Accessions]
  match.name <- names(match.sequence)
  GeneName <- sapply(
    1:length(match.name), function(k) {
      GeneName <- array()
      if(length(grep("GN=", match.name[k]))==0){GeneName[k] <- NA} #2代表蛋白Description所在列数
      else{d <- regexpr("GN=", match.name[k]) #查找在data_ori_matrix[i,2]中第一次出现“GN=”的位置
      e <- substring(match.name[k],d+3) #若找到“GN=”，在第一次出现“GN=”的位置将其及其前面的字符切割
      f <- regexpr(" ",e) #查找在e中第一次出现“ ”的位置
      GeneName[k] <- substring(e, 0, f-1)} #在第一次出现“ ”的位置将其及其前面的字符切割
      GeneName[k] <- gsub("/.*$","",GeneName[k])
    }
  )
  match.df <- data.frame(protein = sapply(strsplit(match.name,"\\|"), function(x) x[2]),
                         GeneName = GeneName,
                         Description = sapply(strsplit(match.name,"\\|"), function(x) x[3]))
  
  data$Gene.Name <- match.df$GeneName[base::match(data$Protein.Accessions,match.df$protein)]
  data$Protein.Description <- match.df$Description[base::match(data$Protein.Accessions,match.df$protein)]
  data <- data[,-c(2,3,4,5)]
  data <- data %>% dplyr::select(Protein.Group,Protein.Accessions,Gene.Name,Protein.Description,everything())
  return(data)
}
protein.data <- process_ids_genes(protein.data)
# 过滤蛋白质（至少在一个组中有50%以上的样本非零）
keep_rows <- apply(protein.data, 1, function(row) {
  all(sapply(group.name, function(group) {
    # group_cols <- grep(paste0("^",group,"\\d+"), colnames(protein.data), value = TRUE)
    group_cols <- which(colnames(protein.data) %in% group.file$SampleName[group.file$Group == group])
    non_zero_count <- sum(!is.na(row[group_cols]))
    non_zero_count < 0.5 * sample_num
  }))
})
data_filtered <- protein.data[!keep_rows, ]
# 使用mixed imputation strategy填补缺失值
## 最小值的二分之一
intensity_matrix <- as.matrix(data_filtered[, -c(1:nonsamplecol)]) # 去除前几列非样本列
min_nonzero_value <- min(intensity_matrix[intensity_matrix > 0], na.rm = TRUE)
half_min_value <- min_nonzero_value / 2
# 定义填补缺失值的函数
fill_na <- function(group_data, half_min_value) {
  col_names <- colnames(group_data)
  # 找到全是 NA 的行并用 half_min_value 填充
  all_na_rows <- apply(group_data, 1, function(row) all(is.na(row)))
  for (name in col_names) {
    group_data[all_na_rows, name] <- half_min_value  + runif(1,0,0.001)
  }
  # 使用 KNN 填充剩余的缺失值
  group_data_filled <- as.data.frame(SeqKNN(as.matrix(group_data), k = 10))
  return(group_data_filled)
}
# 对每个组进行填充
protein_data_impute_group <- lapply(group.name, function(group) {
  # group_indices <- grep(pattern = paste0("^", group, "\\d+"), x = colnames(data_filtered))
  group_indices <- which(colnames(data_filtered) %in% group.file$SampleName[group.file$Group == group])
  group_intensity <- data_filtered[, group_indices]
  protein_data_group <- fill_na(group_intensity, half_min_value)
  return(protein_data_group)
})
# 检查colnames(data_filtered)是否包含"QC"
if (any(grepl("QC", colnames(data_filtered)))) {
  # 获取QC列的索引
  qc_indices <- grep(pattern = "QC", x = colnames(data_filtered))
  qc_intensity <- data_filtered[, qc_indices]
  qc_data_impute <- fill_na(qc_intensity, half_min_value)
  
  # 合并填补后的数据
  protein_data_impute_merge <- do.call(cbind, protein_data_impute_group)
  protein_data_impute <- data.frame(data_filtered[, c(1:nonsamplecol)], protein_data_impute_merge)
  protein_data_impute_addqc <- data.frame(data_filtered[, c(1:nonsamplecol)], protein_data_impute_merge, qc_data_impute)
  pca.data <- protein_data_impute_addqc[, c(2, 5:ncol(protein_data_impute_addqc))]
} else {
  # 如果不包含"QC"，则仅合并填补后的数据
  protein_data_impute_merge <- do.call(cbind, protein_data_impute_group)
  protein_data_impute <- data.frame(data_filtered[, c(1:nonsamplecol)], protein_data_impute_merge)
  pca.data <- protein_data_impute[, c(2, 5:ncol(protein_data_impute))]
}
write.xlsx(pca.data, "temp/pca_data.xlsx")
## 按照分组计算 Mean
for (i in 1:length(group.name)) {
  group.col <- paste0("Mean_",group.name[i])
  protein_data_impute <- protein_data_impute %>% 
    mutate(!!group.col := rowMeans(dplyr::select(., which(colnames(protein_data_impute) %in% group.file$SampleName[group.file$Group == group.name[i]])), na.rm = TRUE))
}
## 计算等电点和覆盖度
all_sequences <- readAAStringSet(fasta_file)
# 提取与UniProt ID匹配的序列信息
all_sequences_name <- sapply(strsplit(names(all_sequences),"\\|"), function(x) x[2])
matched_sequences <- all_sequences[all_sequences_name %in% data_filtered$Protein.Accessions]
protein.fasta <- data.frame(
  UniProt_ID = sapply(strsplit(names(matched_sequences),"\\|"), function(x) x[2]),
  Sequence = as.character(matched_sequences)
)
protein.fasta$pI <- sapply(protein.fasta$Sequence, function(seq) {
  Peptides::pI(seq)
})
## 计算蛋白的覆盖度coverage
peptide.data$Protein.Group <- sapply(strsplit(peptide.data$Protein.Group,";"),FUN = function(x) x[1])
peptide.data.select <- peptide.data[c("Protein.Group","Stripped.Sequence")]
matched_peptide_sequences <- all_sequences[all_sequences_name %in% peptide.data.select$Protein.Group]
peptide.fasta <- data.frame(
  UniProt_ID = sapply(strsplit(names(matched_peptide_sequences),"\\|"), function(x) x[2]),
  Sequence = as.character(matched_peptide_sequences)
)
protein.coverage <- left_join(x = peptide.data.select,y = peptide.fasta,by = c("Protein.Group" = "UniProt_ID"))
protein.coverage <- protein.coverage %>%
  filter(!is.na(Sequence) & Sequence != "", !is.na(Stripped.Sequence) & Stripped.Sequence != "")
# 定义计算覆盖度的函数
calculate_coverage <- function(protein_sequence, detected_peptides) {
  protein_length <- nchar(protein_sequence)
  if (protein_length == 0) {
    return(NA)
  }
  coverage_map <- rep(0, protein_length)
  for (peptide in detected_peptides) {
    if (nchar(peptide) == 0) {
      next
    }
    start <- unlist(gregexpr(peptide, protein_sequence, fixed = TRUE))[[1]]
    if (start != -1) {
      indices <- start:(start + nchar(peptide) - 1)
      coverage_map[indices] <- 1
    }
  }
  coverage <- sum(coverage_map) / protein_length * 100
  return(coverage)
}
# 处理数据并计算覆盖度
coverage_results <- protein.coverage %>%
  group_by(Protein.Group) %>%
  summarise(
    Coverage = calculate_coverage(
      unique(Sequence), 
      unique(Stripped.Sequence)
    )
  )
# 匹配Uniprot注释并保存结果
annotate_proteins <- function(data, uniprot_file, kinase_file, organism, output_file) {
  uniprot_db <- read.xlsx(uniprot_file)
  uniprot_db_filter <- uniprot_db %>% dplyr::select(Entry, Length, Mass,`Gene.Ontology.(GO)`, KEGG)
  data_annotated <- left_join(data, uniprot_db_filter, by = c("Protein.Accessions" = "Entry"))
  new_col_order <- c(
    "Protein.Group", "Protein.Accessions", "Gene.Name", "Protein.Description", 
    "Length", "Mass", 
    setdiff(colnames(data_annotated), c("Protein.Group", "Protein.Accessions", "Gene.Name", "Protein.Description", "Length", "Mass"))
  )
  data_annotated <- data_annotated[, new_col_order]
  colnames(data_annotated)[which(colnames(data_annotated) == "Mass")] <- "Molecular Weight[KDa]"
  data_annotated$`Molecular Weight[KDa]` <- as.numeric(data_annotated$`Molecular Weight[KDa]`) / 1000
  data_annotated$Length <- as.numeric(data_annotated$Length)
  # kegg:K-number,pathway_id
  kegg_db <- read.csv(kegg_file)
  kegg_db$GeneID <- paste(organism_code,kegg_db$GeneID,sep=":")
  data_annotated$KO.ID <- sapply(data_annotated$KEGG,function(x){
    paste(kegg_db$KO.ID[which(kegg_db$GeneID %in% unlist(strsplit(x,";")))],collapse = ";")
  })
  data_annotated$pathway.id <- sapply(data_annotated$KEGG,function(x){
    paste(kegg_db$pathway.id[which(kegg_db$GeneID %in% unlist(strsplit(x,";")))],collapse = ";")
  })
  #interpro
  domain_db <- read.delim2(interpro_file)
  colnames(domain_db)<-c("Protein_ID","MD5.disest","length"	,"source.database","database.ID","Description_1","start","stop",	"e-value"	,"status"	,"Date",	"ID","Description"	,"GO","Pathway")
  domain_db$Protein_ID <- sapply(strsplit(domain_db$Protein_ID,split = "\\|"),FUN = function(x)x[2])
  domain_db <- domain_db %>%
    select(ID, Description, Protein_ID,source.database) %>%
    filter(ID != "-")
  data_annotated$InterPro.id <- sapply(data_annotated$Protein.Accessions,function(x){
    paste(domain_db$ID[which(domain_db$Protein_ID %in% x)],collapse = ";")
  })
  # Subcellular database
  subdb <- read.csv(Sys.glob(paste0(subcellular_file_path, organism_Scientific_name, "/*csv")))
  subdb$UniprotID <- sapply(strsplit(subdb$Protein_ID,split = "\\|"),FUN = function(x)x[2])
  sub.db <- dplyr::select(subdb,UniprotID,Localizations)
  sub.db$Localizations <- sapply(strsplit(sub.db$Localizations,split = "\\|"),FUN = function(x)x[1])
  data_annotated$subcellular_location <- sub.db$Localizations[base::match(data_annotated$Protein.Accessions,sub.db$UniprotID)]
  # Kinase database
  kinase.database <- read.xlsx(kinase_file,startRow = 3)
  kinase.db <- read.xlsx(Sys.glob(paste0(kinase_file_path,organism_category,"/",organism_code,"_",organism_ID,"_Kinase.xlsx")))
  if (organism %in% kinase.database$KIN_ORGANISM) {
    kinase.prot <- unique(kinase.database[kinase.database$KIN_ORGANISM == organism,"KIN_ACC_ID"])
    kinase.prot.total <- c(kinase.prot,kinase.db$Entry)
  }else{
    kinase.prot.total <- kinase.db$Entry
  }
  data_annotated$kinase <- ifelse(test = data_annotated$Protein.Accessions %in% kinase.prot.total,yes = "TRUE",no = "FALSE")
  # TF database
  TFdb <- read.table(Sys.glob(paste0(TF_file_path,organism,"/","*_TF.txt")),header = T,sep = "\t")
  TFdb <- TFdb[,c("Symbol","Family")]
  data_annotated$TF <- TFdb$Family[base::match(data_annotated$Gene.Name,TFdb$Symbol)]
  # DO database
  DOdb <- read.delim("/PERSONALBIO1/prote/database/DO_db/DISEASE-ALLIANCE_COMBINED_6.tsv",skip = 15)
  organism_Scientific_name1<-gsub("_"," ",organism_Scientific_name)
  DOdb <- DOdb %>% filter(SpeciesName == organism_Scientific_name1,DBobjectType == "gene")%>%
    select(DOID,DBObjectSymbol)
  data_annotated$DO <- DOdb$DOID[base::match(data_annotated$Gene.Name,DOdb$DBObjectSymbol)]
  # Wikipathway database
  Wikipathway.db <- read.delim(Sys.glob(paste0(WikiPathway_file_path,organism_Scientific_name,".xls")))
  Wikipathway.db <- Wikipathway.db[,c("wpid","Entry")]
  Wikipathway.db.long <- Wikipathway.db %>%
    group_by(Entry) %>%
    summarise(wpid = paste(wpid, collapse = ";"), .groups = 'drop')
  data_annotated <- data_annotated %>%
    left_join(Wikipathway.db.long, by = c("Protein.Accessions" = "Entry"))
  colnames(data_annotated)[which(names(data_annotated) == "wpid")] <- "Wikipathway"
  
  matched_indices <- match(data_annotated$Protein.Accessions,protein.fasta$UniProt_ID)
  data_annotated$Isoelectric.Point <- protein.fasta$pI[matched_indices]
  data_annotated <- data_annotated %>%
    dplyr::select(1:6, Isoelectric.Point, everything())
  matched_indices_coverage <- match(data_annotated$Protein.Accessions,coverage_results$Protein.Group)
  data_annotated$Coverage <- coverage_results$Coverage[matched_indices_coverage]
  data_annotated <- data_annotated %>%
    dplyr::select(1:7, Coverage, everything())
  
  write.xlsx(data_annotated, file = output_file)
}
annotate_proteins_diff <- function(data, uniprot_file, kinase_file, organism, output_file) {
  uniprot_db <- read.xlsx(uniprot_file)
  uniprot_db_filter <- uniprot_db %>% dplyr::select(Entry,`Gene.Ontology.(GO)`, KEGG)
  data_annotated <- left_join(data, uniprot_db_filter, by = c("Protein.Accessions" = "Entry"))
  new_col_order <- c(
    "Protein.Group", "Protein.Accessions", "Gene.Name", "Protein.Description", 
    setdiff(colnames(data_annotated), c("Protein.Group", "Protein.Accessions", "Gene.Name", "Protein.Description"))
  )
  data_annotated <- data_annotated[, new_col_order]
  # kegg:K-number,pathway_id
  kegg_db <- read.csv(kegg_file)
  kegg_db$GeneID <- paste(organism_code,kegg_db$GeneID,sep=":")
  data_annotated$KO.ID <- sapply(data_annotated$KEGG,function(x){
    paste(kegg_db$KO.ID[which(kegg_db$GeneID %in% unlist(strsplit(x,";")))],collapse = ";")
  })
  data_annotated$pathway.id <- sapply(data_annotated$KEGG,function(x){
    paste(kegg_db$pathway.id[which(kegg_db$GeneID %in% unlist(strsplit(x,";")))],collapse = ";")
  })
  #interpro
  domain_db <- read.delim2(interpro_file)
  colnames(domain_db)<-c("Protein_ID","MD5.disest","length"	,"source.database","database.ID","Description_1","start","stop",	"e-value"	,"status"	,"Date",	"ID","Description"	,"GO","Pathway")
  domain_db <- domain_db %>%
    select(ID, Description, Protein_ID,source.database) %>%
    filter(ID != "-")
  data_annotated$InterPro.id <- sapply(data_annotated$Protein.Accessions,function(x){
    paste(domain_db$ID[which(domain_db$Protein_ID %in% x)],collapse = ";")
  })
  # Subcellular database
  subdb <- read.csv(Sys.glob(paste0(subcellular_file_path, organism_Scientific_name, "/*csv")))
  subdb$UniprotID <- sapply(strsplit(subdb$Protein_ID,split = "\\|"),FUN = function(x)x[2])
  sub.db <- dplyr::select(subdb,UniprotID,Localizations)
  sub.db$Localizations <- sapply(strsplit(sub.db$Localizations,split = "\\|"),FUN = function(x)x[1])
  data_annotated$subcellular_location <- sub.db$Localizations[base::match(data_annotated$Protein.Accessions,sub.db$UniprotID)]
  # kinase database
  kinase.database <- read.xlsx(kinase_file,startRow = 3)
  kinase.db <- read.xlsx(Sys.glob(paste0(kinase_file_path,organism_category,"/",organism_code,"_",organism_ID,"_Kinase.xlsx")))
  if (organism %in% kinase.database$KIN_ORGANISM) {
    kinase.prot <- unique(kinase.database[kinase.database$KIN_ORGANISM == organism,"KIN_ACC_ID"])
    kinase.prot.total <- c(kinase.prot,kinase.db$Entry)
  }else{
    kinase.prot.total <- kinase.db$Entry
  }
  data_annotated$kinase <- ifelse(test = data_annotated$Protein.Accessions %in% kinase.prot.total,yes = "TRUE",no = "FALSE")
  # TF database
  TFdb <- read.table(Sys.glob(paste0(TF_file_path,organism,"/","*_TF.txt")),header = T,sep = "\t")
  TFdb <- TFdb[,c("Symbol","Family")]
  data_annotated$TF <- TFdb$Family[base::match(data_annotated$Gene.Name,TFdb$Symbol)]
  # DO database
  DOdb <- read.delim("/PERSONALBIO1/prote/database/DO_db/DISEASE-ALLIANCE_COMBINED_6.tsv",skip = 15)
  organism_Scientific_name1<-gsub("_"," ",organism_Scientific_name)
  DOdb <- DOdb %>% filter(SpeciesName == organism_Scientific_name1,DBobjectType == "gene")%>%
    select(DOID,DBObjectSymbol)
  data_annotated$DO <- DOdb$DOID[base::match(data_annotated$Gene.Name,DOdb$DBObjectSymbol)]
  # Wikipathway database
  Wikipathway.db <- read.delim(Sys.glob(paste0(WikiPathway_file_path,organism_Scientific_name,".xls")))
  Wikipathway.db <- Wikipathway.db[,c("wpid","Entry")]
  Wikipathway.db.long <- Wikipathway.db %>%
    group_by(Entry) %>%
    summarise(wpid = paste(wpid, collapse = ";"), .groups = 'drop')
  data_annotated <- data_annotated %>%
    left_join(Wikipathway.db.long, by = c("Protein.Accessions" = "Entry"))
  colnames(data_annotated)[which(names(data_annotated) == "wpid")] <- "Wikipathway"
  write.xlsx(data_annotated, file = output_file)
}
## total proteins
## 按照对比策略计算FC log2FC pvalue Qvalue
protein_data_impute.total <- list()
for (j in 1:length(compare.name)) {
  print(compare.name[j])
  compare.group <- unlist(strsplit(x = compare.name[j],split = "_vs_"))
  mean1 <- paste0("Mean_",compare.group[1])
  mean2 <- paste0("Mean_",compare.group[2])
  # 计算 Fold Change (FC) 和 Log2 Fold Change (Log2FC)
  protein_data_impute <- protein_data_impute %>%
    mutate(FC := .[[mean2]] / .[[mean1]],
           Log2fc := log2(.[[mean2]] / .[[mean1]]))
  # subgroup1 <- grep(paste0("^",compare.group[1],"\\d+"), colnames(protein_data_impute), value = TRUE)
  # subgroup2 <- grep(paste0("^",compare.group[2],"\\d+"), colnames(protein_data_impute), value = TRUE)
  subgroup1 <- colnames(protein_data_impute)[which(colnames(protein_data_impute) %in% group.file$SampleName[group.file$Group == compare.group[1]])]
  subgroup2 <- colnames(protein_data_impute)[which(colnames(protein_data_impute) %in% group.file$SampleName[group.file$Group == compare.group[2]])]
  df_a <- protein_data_impute[which(colnames(protein_data_impute) %in% subgroup1)]
  df_b <- protein_data_impute[which(colnames(protein_data_impute) %in% subgroup2)]
  protein_data_impute$p_value <- sapply(1:nrow(df_a), function(k) {
    col_a <- as.numeric(df_a[k,])
    col_b <- as.numeric(df_b[k,])
    if(mean(col_b) == mean(col_a)) {
      return(1)
    } else {
      var_equal <- var.test(col_a, col_b)$p.value > 0.05
      return(t.test(col_a, col_b, var.equal = var_equal)$p.value)
    }
  })
  protein_data_impute$Q_value <- p.adjust(protein_data_impute$p_value, method = "BH", 
                                          n = length(protein_data_impute$p_value))
  protein_data_impute$Regulation <- ifelse(protein_data_impute$p_value > dif_p, "Nodiff",
                                           ifelse(protein_data_impute$FC > foldchange1, "Up",
                                                  ifelse(protein_data_impute$FC < foldchange2, "Down", "Nodiff")))
  colnames(protein_data_impute)[which(colnames(protein_data_impute) == "FC")] <- paste0(compare.group[1],"_vs_",compare.group[2],".FC")
  colnames(protein_data_impute)[which(colnames(protein_data_impute) == "Log2fc")] <- paste0(compare.group[1],"_vs_",compare.group[2],".Log2fc")
  colnames(protein_data_impute)[which(colnames(protein_data_impute) == "p_value")] <- paste0(compare.group[1],"_vs_",compare.group[2],".pvalue")
  colnames(protein_data_impute)[which(colnames(protein_data_impute) == "Q_value")] <- paste0(compare.group[1],"_vs_",compare.group[2],".Q_value")
  colnames(protein_data_impute)[which(colnames(protein_data_impute) == "Regulation")] <- paste0(compare.group[1],"_vs_",compare.group[2],".Regulation")
  protein_data_impute.total[[j]] <- protein_data_impute
  ## diff prot
  diff.group.col <- c(which(colnames(protein_data_impute) %in% c(mean1,mean2,subgroup1,subgroup2)),
                      grep(pattern = compare.name[j],x = colnames(protein_data_impute)))
  protein_data_impute_diff <- protein_data_impute[,c(1:nonsamplecol,diff.group.col)]
  protein_data_impute_diff_reg <- protein_data_impute_diff[which(colnames(protein_data_impute_diff) == paste0(compare.group[1],"_vs_",compare.group[2],".Regulation"))]
  protein_data_impute_diff_final <- protein_data_impute_diff[protein_data_impute_diff_reg != "Nodiff",]
  annotate_proteins_diff(protein_data_impute_diff_final, uniprot_file, kinase_file, organism, paste0(output2,compare.name[j],"-DEP_results.xlsx"))
  annotate_proteins_diff(protein_data_impute_diff, uniprot_file, kinase_file, organism, paste0("temp/",compare.name[j],"-total_results.xlsx"))
}
protein_data_impute.total.merge <- do.call(cbind, protein_data_impute.total)
duplicated_cols <- duplicated(colnames(protein_data_impute.total.merge))
protein_data_impute.total.final <- protein_data_impute.total.merge[, !duplicated_cols]
annotate_proteins(protein_data_impute.total.final, uniprot_file, kinase_file, organism, paste0(output1,"Protein_SummaryStat.xlsx"))
